requirements.txt
Django
feincms
djangorestframework
django-cors-headers
django-extensions
easy_thumbnails
psycopg2-binary
django-filter
python-app-loader
six